libseccomp: initial import of project
This patch is the initial import of the libseccomp library, a library
intended for use by application developers who wish to leverage the
enhanced seccomp (mode 2) support in the Linux Kernel but who do not
want to craft their own seccomp filter code by hand.  This library will
provide a high-level API for developers which will allow them to
generate and load seccomp filter code into the kernel by specifying
which syscalls, optionally with argument values, are to be allowed or
denied by the kernel.

Needless to say, this is an initial import into the git repository so
everything is still subject to change, and the usual caveats about
untested code apply quite strongly.

This code is licensed under the GPLv2, with some portions taken from
examples provided by Will Drewry <wad@chromium.org> and copyrighted
by the Chromium OS Authors <chromium-os-dev@chromium.org> which carries
the following license (included in the associated files as well):

 "The code may be used by anyone for any purpose, and can serve as a
  starting point for developing applications using
  prctl(PR_ATTACH_SECCOMP_FILTER).
  No guarantees are provided with respect to the correctness or
  functionality of this code."

Signed-off-by: Paul Moore <pmoore@redhat.com>