gen: fix some problems with label handling, do some cleanup, and minor optimization
We currently store labels on the stack, which is a big no-no as the BPF
helper functions do not makes copies of the label strings which result
in lots of garbage in the labels when you try to resolve them.  Fun.

Thanks to Eric Paris for finding this problem and providing a partial
fix.

This patch creates a label store which is used as a stable place to
maintain the labels while we are creating the BPF representation.  In
addition, we do a good deal of cleanup to the BPF code (still needs a
whole lot more), and another simple optimization as shown below:

The current code generates something like the following:

#
# filter pseudo code start
#
# filter code for syscall #1
 if (syscall != 1) goto syscall_1_end;
 if (a0 != 0x1) goto syscall_1_c0_next;
 action ALLOW;
 syscall_1_c0_next:
 if (a0 != 0x2) goto syscall_1_c1_next;
 action ALLOW;
 syscall_1_c1_next:
 if (a1 == 0x0) goto syscall_1_c2_next;
 action ALLOW;
 syscall_1_c2_next:
 syscall_1_end:
# default action
 action DENY;
#
# filter pseudo code end
#

This patch changes the generated code to look like the following:

#
# filter pseudo code start
#
# filter code for syscall #1
 if (syscall != 1) goto syscall_1_end;
 if (a0 != 0x1) goto syscall_1_c0_next;
 action ALLOW;
 syscall_1_c0_next:
 if (a0 != 0x2) goto syscall_1_c1_next;
 action ALLOW;
 syscall_1_c1_next:
 if (a1 == 0x0) goto syscall_1_end;
 action ALLOW;
 syscall_1_end:
# default action
 action DENY;
#
# filter pseudo code end
#

Signed-off-by: Paul Moore <pmoore@redhat.com>