tools: a very basic BPF disassembler for use in development/debugging
An initial take at a BPF disassembler for use is development and
debugging of the BPF programs generated by libseccomp.  There is
still much of the BPF language that is not decoded, but this
initial version should support all of the opcodes that are currently
used in libseccomp's BPF generator.

Signed-off-by: Paul Moore <pmoore@redhat.com>