bpf: fix buffer allocation calculation
We were passing and storing a value called 'len' that was actually the
number of blocks, not the allocation length.  We then were doing
comparisons of the number of blocks to the size of an allocation.

Rename the 'len' variable to 'num_blks' and fix the comparison to always
use allocation lengths.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>