bpf/db: new BPF and filter DB implementations
While the existing filter DB worked, the same could not be said for
the existing BPF generator: it was broken.  When I say broken, I
don't mean it produced really inefficient BPF - it did - I mean it
produced incorrect BPF; take a closer look at the accumulator's
state and how we check syscall numbers to see what I mean.  In order
to solve these problems, both the incorrect BPF and the terribly
inefficient BPF, I offer this ridiculously large patch that I would
never accept in a more "mature" project, but considering where we
are at right now, I'll consider looking the other way ...

Here is the general idea behind this new approach:

 * Filter DB

When new filter rules are added, they are normalized as much as
possible (to make duplicate filter chains easier to identify) and
inserted into a sort of filter/decision tree/graph which allows us
to easily identify and merge duplicate portions at the top of the
filter tree.  Also, with the exception of the top level syscall
entries, we partially sort/order the filter tree in a manner which
helps up later when we generate the BPF; we attempt to order the
tree in such a way that it allows us to preserve the BPF accumulator
state as much as possible (reducing the number of BPF load
instructions required).

 * BPF Generator

The BPF generator takes the filter DB and generates a series of BPF
instruction blocks which it then de-duplicates, links, and uses to
generate the final BPF filter program.  The core idea behind these
BPF instruction blocks is that each block does not require a certain
BPF machine state to execute, each block resets the accumulator (if
needed), and the only persistent memory storage is the syscall/args
record (scratch memory is not considered persistent from a block's
point of view).  The advantage to this is that if we decompose the
entire BPF filter program into blocks, we can order the blocks
however we like and still generate a working BPF filter program.

We manage relationships and jumps between blocks through a hash
values.  After each block has been generated we run a hash function
over the block and use it as a unique identifier (hash includes the
jump targets since we start with the "leaf" blocks) for the block.
At this same point, we can use the hash value as a way of identifying
and removing duplicate blocks without causing damage to the BPF
filter; the new filter DB implementation is good at reducing
duplication at the top of the filter tree, but duplication can still
result towards the bottom of the tree.  Finally, starting at the top
level of the filter tree, we build up a linearly transposed BPF
filter program using the blocks, resolving the jumps based on hash
values as we go along.

Thanks to Eric Paris and his topological sorting idea for starting
me down the path of a better BPF generator and to Corey Bryant for
providing some comments which helped improve the filter DB.

Patch History:
 * v2
   - Fixed the whitespace damage
   - Improved the "fall-through" case so that we jump to the default
     action instead of generating a new return instruction
   - Filled in some memory management todos (free/cleanup)
   - Incorporated comments from Corey
 * v1 (February 9th, 2012)
   - Initial version, generates correct BPF for 02-basic-pfc like
     tests
   - Only supports 32-bit systems at present
   - Lacking important and necessary comments
   - Lacking proper memory management and cleanup
   - Jump distance verification and long jump insertion (worst case)
     areas have been identified but are marked as TBD/XXX
   - The final assembly algorithm is simplistic at best, we can do
     much better
   - In need of some general cleanup

Signed-off-by: Paul Moore <pmoore@redhat.com>