db/bpf: fix some bugs found in testing
Ashley found a problem with the following filter configuration:

 seccomp_add_syscall(SCMP_ACT_ALLOW, SCMP_SYS(read), 1,
                     1, SCMP_CMP_EQ, buf);
 seccomp_add_syscall(SCMP_ACT_ALLOW, SCMP_SYS(read), 2,
                     0, SCMP_CMP_EQ, 17,
                     1, SCMP_CMP_EQ, buf);

At present the code would only generate BPF/PFC for the second call
to seccomp_add_syscall(), ignoring the first.  We could ignore the
second call, as an optimization, but we must honor the first call to
ensure correctness.

This patch fixes a problem in db_add_syscall() to correct a problem
where new argument chains were not being added correctly, and a few
other bugs which were exposed as a result of this change.

Reported-by: Ashley Lai <adlai@linux.vnet.ibm.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>