tools: BPF simulator
A very simple BPF simulator designed to make it easier to debug and
do automated testing of long and complicated BPF programs generated
by libseccomp.

This is an initial version with only a small subset of supported BPF
commands, however, it should be sufficient to simulate most, if not
all of the BPF programs generated by the current version of
libseccomp.

Signed-off-by: Paul Moore <pmoore@redhat.com>