gen_bpf: assign whole static bpf_jump structs rather than individual parts
Inside the _BPF_INSTR macro we were assigning the bpf_jump portions as
separate parts instead of as a whole.  This meant that if we ever change
the bpf_jump structure we might not be doing the assignment correctly.
I also don't know the semantics of setting one part of a union and
assigning another part.  I know that the void * is the largest part of
this union, but what if we added a larger thing to the union?  Would
just assigning the void * be enough?  I dunno.  This way it's the
compiler's problem.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>