bpf: remove needless absolute jump at end of tg_sys rules
The tg_sys blocks are all syscalls with no chain behind them.  On true
they always jump to the default and on false they always go to the next
line.  At the end of the block we put an absolute jump by one line to
skip the default result.  Since we know jf is always 0 we can just munge
the jf on the last entry to do the jump.  This saves one line of BPF and
should save one conditional on a potentially hot path in the BPF.

Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>