bpf: make the syscall-only case less of a special case
Stop treating the syscall-only, no chain, case as a special case.
This not only simplifies the code a great deal, it also makes it
easier to implement improved/optimized filter ordering as we can
now treat all of the syscall filters the same way.

This patch also augments the 04-chains-bpf test.

Signed-off-by: Paul Moore <pmoore@redhat.com>