db/bpf: add the notion of rule priority
The basic idea is to move the filter rules with a smaller number of
checks (easier/quicker to evaluate) closer to the top of the of the
BPF filter code.  This is a more generalized approach to the current
approach of placing filters without chains at the top of the
generated BPF code.

This also paves the way for allowing user hints to further influence
filter rule ordering in the final BPF.

Signed-off-by: Paul Moore <pmoore@redhat.com>