bpf: remove some old comments
While the removed comments are still true in the strictest sense,
we do do a reasonable job of ordering blocks now and while there is
a lot of iterating over the instruction blocks, it happens for a
good reason.

In short, we can probably do these things better, but right now they
are no worse than the rest of the code.

Signed-off-by: Paul Moore <pmoore@redhat.com>