tests: Consolidate BPF/PFC getopt() test code
Each of the existing tests has getopt() code to determine whether
the test output should be bpf or pfc.  This patch consolidates that
code into a common src/obj.  A new utility file is introduced to
contain this code.  For now the utility file only contains the
bpf/pfc getopt() code, but can be expanded with additional common
test code in the future.

Signed-off-by: Corey Bryant <coreyb@linux.vnet.ibm.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>