bpf: resolve issues with duplicate instruction blocks in complex filters
We had a problem where sometimes we would free an instruction block
when adding it to the hash table (if it was a duplicate block) and
later try to reference the free'd instruction block via another
instruction block's jump which is very not good.

This patch fixes this by storing the duplicate blocks instead of
freeing them.  It results in a slightly larger memory footprint but
prevents us walking free'd memory.

Signed-off-by: Paul Moore <pmoore@redhat.com>