bpf: manage jump lengths to return instruction blocks
The ability for callers to specify syscall priorities has the
potential to result in some very long jumps to a simple return
instruction (the syscall only, zero argument chain rules may not
always be grouped together anymore).  This function attempts to
resolve this by adding extra return instructions when the jump
length to an instruction block grows too large.

It also has the benefit of solving the optimization problem
where we insert a long jump to jump to a return instruction when
we could have simply duplicated the return instruction.

At present this patch sets the maximum return instruction jump
length to the maximum possible value - 255 - we can play with
reducing this later.  Decreasing the value will result in shorter
jumps but possibly longer filter code; increasing the value will
result in shorter filter code but longer jumps.

	#define _BPF_JMP_MAX_RET	255

Signed-off-by: Paul Moore <pmoore@redhat.com>