db: optimize for 64bit systems
Finally optimize the filter for 64bit systems by make the filter DB
layer much more aware of the fact that we need to use a 32bit
machine (the BPF machine in the kernel) to evaluate 64bit values.
The basic change is to move the 64bit-to-32bit logic from the BPF
layer to the filter DB layer; as one would expect, it greatly
simplifies the BPF layer but makes the filter DB more complex.  The
PFC output also suffers a bit in terms of readability, but the new
BPF more than makes up for the PFC loss.  Granted it is an extreme
case, but the 05-long-jumps tests goes from 3031 BPF instructions
before this patch to 1227 BPF instructions (!) with this patch
applied.

This patch also adjusts the AINC_BLK value to 2 as it appears to be
the largest instruction block allocation size we need at present
and helps shrink our memory footprint.

Signed-off-by: Paul Moore <pmoore@redhat.com>