api: tweak the rule_add API further
This patch makes seccomp_rule_add() operate slightly differently
which should make it more useful form an arch independence point of
view; if a filter rule needs to be rewritten for a particular arch
and the full rule can not be preserved, the function no longer
fails.  While this may be less secure in the stricktest sense, it
should make the library much more usable.

Also, the seccomp_rule_add_exact() function was added with the older
behavior.

Signed-off-by: Paul Moore <pmoore@redhat.com>