arch: enable less strict filter rewrites
At present we fail if we can't completely preserve the caller's
filter, while admirable, this does require some knowledge of the
architecture to ensure you're adding a "correct" rule.

In keeping with our goal of architecture independence, we want to
add the ability to do "best effort" rewrites that preserve as much
of the original filter rule as possible.

Signed-off-by: Paul Moore <pmoore@redhat.com>