api: make the va_args more reliable
While va_args are nice, testing has proven them to be unreliable
with respect to datum types and this is causing some real problems
on 64 bit architectures.

This patch resolves this by moving the individual argument, op, and
datum values into a structure that can be created by the SCMP_CMP()
and SCMP_A{0-6}() macros.

Signed-off-by: Paul Moore <pmoore@redhat.com>