tools: fix bugs for bpf_sim seccomp_data indexing and acc tests
This patch fixes a few bugs in bpf_sim.

We were previously only storing a byte of the seccomp_data structure
in the accumulator.  This is updated to store the full word of the
corresponding nr, arch, or arg in the accumulator.

Also the relational tests of the accumular value vs filter value were
backwards and fixed in this patch.

A missing break is also added in case '4' of the main switch statement.

Signed-off-by: Corey Bryant <coreyb@linux.vnet.ibm.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>