bpf: fix a problem when failing on the low word check on 64bit systems
When checking the low word on a 64bit system we would incorrectly
jump to the default action if there were no other argument checks
on the current level when the correct behavior is to back up a level
(to the high word check) and continue from there.

This patch fixes this behavior and should restore proper filter
generation on 64bit systems.

Reported-by: Corey Bryant <coreyb@linux.vnet.ibm.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>