tests: Add bpf-sim-fuzz regression test type
This patch adds the bpf-sim-fuzz regression test type and corresponding
test data to each of the existing batch files.  For more information
on the bpf-sim-fuzz test type, please refer to the
run_test_bpf_sim_fuzz() function comments.

Also included in this patch is a fix for a potential bug in the printf
width that was being used for the final column of log file test data
output for the bpf-sim test type.  It was previously indexing beyond
the bounds of the COL_WIDTH array.  This has been fixed to not use a
column width for the final column since it is not necessary.

Signed-off-by: Corey Bryant <coreyb@linux.vnet.ibm.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>