api: rename the seccomp_gen_*() functions to seccomp_export_*()
These functions were originally intended as a debugging tool for
developers (both application and libseccomp), but recent discussions
have led me to believe that at some point in the future libseccomp
will likely support the import/export of seccomp policy.  While I
hate changing the API, we haven't released libseccomp yet so we are
still free to make what changes we need and I think this rename puts
us in a better position for the future.

Signed-off-by: Paul Moore <pmoore@redhat.com>