api: remove the internal state and make the API context-aware
IMPORTANT: WILL CAUSE BREAKAGE WITH v0.1.x API

The current API relies on the library storing internal state about
the filter context.  While this is okay in several cases, it may
cause problems in others, e.g. threaded applications.

Since the bulk of the library already operates on a filter context,
known as "struct db_filter", this patch simply adds an additional
parameter to the public API, exporting this context as an opaque
context/handle.

Signed-off-by: Paul Moore <pmoore@redhat.com>