build: fix some pkgconfig problems during configure/install
libseccomp should not assume the existence of the prefix and library
directory since it can be an empty directory used for testing.
The best we can do is check if the destination path is not a directory
if it exists.

Added a macro for installing .pc files that creates a directory if it
doesn't exists and copy the file to the destination, respecting
ownership with proper file permissions.

Signed-off-by: Thiago Marcos P. Santos <thiago.santos@intel.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>