db: enable basic filter collection support
In order to support systems that can run applications from multiple
architectures we need to be able to support multiple filter DBs; were
calling this "filter collections".  This patch adds the basic
collection support such that it passes all of the existing tests;
further work may be necessary once we start using the multiple filter
capabilities.

Signed-off-by: Paul Moore <pmoore@redhat.com>