tools: make the sys_resolver tool work for non-native architectures
In order to test non-native architectures using our existing regression
test infrastructure and BPF simulator we need to be able to resolve
syscall names on arbitrary architectures.

This patch moves aside the existing bash script and replaces it with a
small C progream which uses some internal libseccomp routines to resolve
the syscall.

Signed-off-by: Paul Moore <pmoore@redhat.com>