bpf: fix a problem when creating filters that include syscall arguments
In some cases when generating BPF filters, if a particular argument chain
did not match, the filter would fall through to the default action for the
architecture instead of checking on other argument chains for the same
syscall.  This patch corrects this problem by ensuring that all argument
chains are evaluated.

Signed-off-by: Paul Moore <pmoore@redhat.com>