tools: allow sys_resolver to translate syscalls
Normally sys_resolver does not translate syscalls which means that in
some cases, e.g. socket() on x86, the returned syscall number could be
a negative number (__PNR_socket).  This patch adds a new option, '-t',
which causes sys_resolver to attempt to do the translation and return
the translated syscall number instead, e.g. socketcall() for socket() on
x86.

Signed-off-by: Paul Moore <pmoore@redhat.com>