python: ensure attributes are treated as 32 bits
Retrieving attributes using the Python bindings fails on some platforms.
The attributes are encoded in a 32-bit mask.  Python variables are
usually larger (64 bits); Cython is not capable of recognizing that it
should only use a 32-bit number on every platform.  This patch ensures
that the variable used to store the value of the attribute is only 32 bits.

Signed-off-by: Michael Strosaker <strosake@linux.vnet.ibm.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>