tests: test the bpf accumulator checking logic
When building the BPF filter code we need to ensure that we take the
state of the BPF state machine accumulator into account.  This test
creates a situation where the BPF filter code generator needs to
perform some extra work to ensure the accumulator state is correct.

This test is based on a bug reproducer by Matthew Heon.

Signed-off-by: Paul Moore <pmoore@redhat.com>