db: optimize masked compares
If the masked compare is a tautology we don't need to generate
instructions for the runtime test. It'll always be true.

This patch handles the case for 32 bit arches and partially for 64 bit
arches. The cases where either the upper half or the lower half is a
tautology is still TODO.

Signed-off-by: Mathias Krause <minipli@googlemail.com>
[PM: minor function name changes to better match existing style]
Signed-off-by: Paul Moore <pmoore@redhat.com>