tests: extend 12-sim-basic_masked_ops with 64 bit values
Add test vectors with bits set in the upper half of the syscall argument.
They trigger a bug with mask values having the upper half set to 0. We
accidentally emit a test for 0 in this case when we should not test the
upper half at all.

Signed-off-by: Mathias Krause <minipli@googlemail.com>
Signed-off-by: Paul Moore <pmoore@redhat.com>
(imported from commit 62f1e7e2fa6163964322d232bde89e37478772e9)