tests: workaround problems with large integers on 32-bit ARM
See the mailing list thread below:

 -> https://groups.google.com/forum/#!topic/libseccomp/VtrClkXxLGA

... unfortunately the 32-bit ARM userspace has problems with this
particular test so we need to disable it for the time being.  It is
important to note that this is only a problem with the test and not
with libseccomp in general.

Signed-off-by: Paul Moore <pmoore@redhat.com>
(imported from commit bca4e115715174a64c7b5f56430a51f3e676c34a)