db: introduce transaction support
This patch adds basic transaction support to the db layer.  The db
transactions allow callers to checkpoint a filter collection at the
current point in time and later rollback the filter collection if
necessary.  There will be some overhead at the start of the
transaction to generate the checkpoint, but transaction commit and
abort rollbacks are guaranteed to succeed.  Transactions may be
nested.

Signed-off-by: Paul Moore <pmoore@redhat.com>