all: use the seccomp() syscall whenever possible for tested ABIs
The seccomp() syscall was first added in Linux 3.17 so most systems
should now support this syscall.  Most importantly, the use of the
seccomp() syscall enabled the thread sync functionality which isn't
possible with prctl(); although callers still need to enable the flag
per-filter as the thread sync default is disabled.

This patch also unified the return values of the sys_chk_*()
functions.

Signed-off-by: Paul Moore <paul@paul-moore.com>