arch: fix a number of 32-bit x86 failures related to socket syscalls
It turns out there was still a few bugs with the 32-bit x86 socket
syscalls, especially on systems with older kernel headers installed.
This patch corrects these problems and perhaps more importantly,
returns the resolver API functions to returning the negative pseudo
syscall numbers in the case of 32-bit x86, this helps ensure things
continue to work as they did before as the API does not change.

It it important to note that libseccomp still generates filter code
for both multiplexed and direct socket syscalls regardless.

Signed-off-by: Paul Moore <paul@paul-moore.com>