tools: Add arch resolver utility
Add a utility, scmp_arch_resolver, to translate architecture tokens to
human readable strings and back. This may be useful, for example, when
needing to understand the "arch=N" portion of seccomp audit messages.

The Makefile.am is configured so that scmp_arch_resolver is not
installed at this time but that's something that can always be changed
if the utility proves to be useful.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>