db: include the arguments in the db_api_rule_list struct
Instead of dynamically allocating a variable number of arguments,
include an array of ARG_COUNT_MAX elements directly in the struct.
Also perform a number of simplifications to the code with the
understanding that ARG_COUNT_MAX is an ABI independent value that
isn't variable.

Signed-off-by: Paul Moore <paul@paul-moore.com>
(imported from commit e3addce3794ddb6dc174d429da055296282df0e6)