all: treat syscall -1 as a valid syscall
Process tracers use a -1 syscall value to indicate that a syscall
should be skipped.  This turns out to be quite an undertaking as
we need to workaround __NR_SCMP_ERROR (which also has a value of
-1).  Pay special attention to the new attribute,
SCMP_FLTATR_API_TSKIP, and the documentation additions.

More information in the GitHub issue:
* https://github.com/seccomp/libseccomp/issues/80

Signed-off-by: Paul Moore <paul@paul-moore.com>
(imported from commit dc879990774b5fe0b5d3362ae592e8a5bb615fbb)