db: fix 64-bit argument comparisons
Our approach to doing 64-bit comparisons using 32-bit operators was
just plain wrong, leading to a number of potential problems with
filters that used the LT, GT, LE, or GE operators.  This patch fixes
this problem and a few other related issues that came to light in
the course of fixing the core problem.

A special thanks to Jann Horn for bringing this problem to our
attention.

Signed-off-by: Paul Moore <paul@paul-moore.com>