api: stop defining __NR_x values for syscalls that don't exist
Historically libseccomp has created a __NR_x definition for every
syscall it supports, even those that aren't valid for a given ABI.
While this seemed like a good idea at the time, it turned out to have
some unwanted and nasty side effects.  This patch finally corrects
this problem.

The basic approach is quite simple: move the SCMP_SYS() macro to use
__SNR_x values instead of __NR_x values.  The unfortunate side effect
of this change is that instead of just worrying about #defines for the
__PNR_x values we now have to have a __SNR_x define for *every*
syscall.  The good news is that after this patch that should only be
a few new syscalls every year - a very manageable task.

Signed-off-by: Paul Moore <paul@paul-moore.com>