ci/cd: use default python instead of nightly
Previously the live python tests failed in the TravisCI environment because the nightly python version changed in a subtle way that make `cython` install unreliably. This change reverts to a stable version of python that is able to install cython reliably. Applied with the previous commit to disable the live python tests in CI (so that only simulated ones run), this should fix CI.

Signed-off-by: Chris Waldon <chris.waldon@ibm.com>