syscalls: use gperf for lookup
use gperf to generate a perfect hash to lookup syscall names.   It
improves significantly the complexity for seccomp_syscall_resolve_name.*
since it replaces the expensive strcmp for each syscall in the
database, with a lookup table.

The complexity for syscall_resolve_num is not changed and it
uses the linear search, that is anyway less expensive than
seccomp_syscall_resolve_name.* as it uses an index for comparison
instead of doing a string comparison.

On my machine, calling 1000 seccomp_syscall_resolve_name_arch and
seccomp_syscall_resolve_num_arch over the entire syscalls DB passed
from ~0.45 sec to ~0.06s.

Closes: https://github.com/seccomp/libseccomp/issues/207

Signed-off-by: Giuseppe Scrivano <gscrivan@redhat.com>